requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
stripe==10.3.0
gunicorn==22.0.0
